Embed client.html into ECU.exe
Use the rc compiler to embed client/client.html into
the executable. There is no need to give user access
to the html and javascript.

If a file "style.css" is in the same director as the exe,
the stylesheet will be loaded.

Added linker option '/RELEASE' to release builds only.
It adds metadata to exe headers, Windbg was complaining
about it missing.