Move CrBrowser to seperate thread
Move the browser message loop away from the main thread.
We don't have to deal with CefDoMessageLoopWork() anymore but we have to
call the startup and shutdown functions. The main thread id is passed to
SimpleApp then is passed to the handler, which calls PostThreadMessage
with WM_CLOSE when the browser window is closed