Switch to Rawinput and g3log
Raw input uses window messages to signal when a registered device
signals some input, because of this we can switch to GetMessage and not
worry about CPU usage. The main message loop will have to be modified to
allow waiting for ibt folder changes

g3log allows for asynchronous thread-safe logging. By default it uses a
file logger but we can include "sinks" to send logs to the OutputDebug
stream and eventually to the Websocket stream