Add Multithreaded sockets remove file exception
Add multithreaded sockets to ECU, all it can do now is send back what it
receives and has a limit of 5 connections. WinMain will create an event
which is signaled when the main thread receives WM_CLOSE. It'll pass it
to ws_start_daemon which is the thread manager for all of the
connections.

When ws_start_daemon detects there's a new connection it will spawn a
new thread and pass it the accepted socket. The new thread also has the
event from earlier. The worker thread will wait for 2 seconds for data
to be received then process it. After the timeout it will check if the
event has been signaled, if so it will free its resources and notify the
main thread that it exited by decrementing a global counter and closing
its thread handle (which is referenced by the main thread). This is all
protected by a Critical Section. The thread manager works the same way,
wait for new connections for 2 seconds, after the timeout check if the
event has been signaled. If so wait for all worker threads to finish
then free all resources and exit. The main thread meanwhile waits for
the thread manager to exit.