Add force feedback support
The force feedback feature doesn't use 360hz telemetry or any filter on the
original signal. It only sends the signal to the ffb device unmodified.

There is minimal loss of latency (0-2 milliseconds), this is due to
performance reasons, cpu usage is under 2% on a i7 2600k. Loss of
latency is prefferable to losing frametime.